feat(playground): Insights page works in the demo
The demo had no /alerts/analytics handler, so the Insights page was empty in
playground mode. Now it runs the REAL analytics engine:

- computeAlertAnalytics moves to @OpsiMate/shared (pure function; the server
  module re-exports it so its import path and tests stay unchanged)
- MSW handler builds deterministic episode histories per mock alert
  (hash-staggered firing→resolved cycles over ~10 weeks, ending in the alert's
  current state) plus human events from the generated base history and live
  sandbox actions, then computes the same aggregates production serves
- Dashboard scoping (filters/search), tag research, and all trends behave
  identically to production because it IS the production code path
