+++ apps/client/src/mocks/handlers.ts
@@ -6,8 +6,11 @@ import {
 	AlertStatus,
 	OncallTeam,
 	applyAlertListQuery,
+	computeAlertAnalytics,
 	computeAlertFacets,
 	computeAlertGroupSummaries,
+	EpisodeRow,
+	UserEventRow,
 } from '@OpsiMate/shared';
 import { http, HttpResponse } from 'msw';
 import { getPlaygroundUser, OncallTeamState, playgroundState, randomId } from './state';
@@ -72,6 +75,50 @@ const generateBaseHistory = (alertId: string): AlertHistoryData[] => {
 			description: 'Alert resolved',
 		},
 	];
+};
+
+// Deterministic episode history per alert for the Insights demo: a few past
+// firing→resolved cycles spread over the last ~10 weeks (hash-staggered durations
+// and gaps so MTTR/MTBF/peak-hours have real variety), ending in the alert's
+// CURRENT state so the analytics agree with what the table shows.
+const generateEpisodeRows = (alert: Alert, active: boolean): EpisodeRow[] => {
+	const h = hashAlertId(alert.id);
+	const now = Date.now();
+	const rows: EpisodeRow[] = [];
+	let cursor = now - (30 + (h % 40)) * DAY;
+	const cycles = 2 + (h % 4);
+	for (let i = 0; i < cycles && cursor < now - 2 * DAY; i++) {
+		const durationMs = (20 + ((h >> (i + 2)) % 420)) * MINUTE;
+		rows.push({ alertId: alert.id, status: 'firing', at: new Date(cursor).toISOString() });
+		rows.push({ alertId: alert.id, status: 'resolved', at: new Date(cursor + durationMs).toISOString() });
+		cursor += (2 + ((h >> i) % 8)) * DAY + (h % 24) * HOUR;
+	}
+	if (active) {
+		rows.push({ alertId: alert.id, status: 'firing', at: alert.startsAt });
+	} else {
+		const start = now - (1 + (h % 5)) * DAY;
+		rows.push({ alertId: alert.id, status: 'firing', at: new Date(start).toISOString() });
+		rows.push({
+			alertId: alert.id,
+			status: 'resolved',
+			at: new Date(start + (30 + (h % 300)) * MINUTE).toISOString(),
+		});
+	}
+	return rows;
+};
+
+// Human actions feeding MTTA/top-responders: the generated base history's actor
+// events plus everything the user did live in the sandbox.
+const analyticsEvents = (alerts: Alert[]): UserEventRow[] => {
+	const rows: UserEventRow[] = [];
+	for (const alert of alerts) {
+		const events = [...generateBaseHistory(alert.id), ...(playgroundState.alertHistoryEvents[alert.id] ?? [])];
+		for (const event of events) {
+			if (!event.actorName) continue;
+			rows.push({ alertId: alert.id, at: event.date, actorName: event.actorName });
+		}
+	}
+	return rows;
 };
 
 // Appends a live event to an alert's history (newest first).
@@ -397,6 +444,54 @@ export const handlers = [
 
 	http.get(`${API_BASE}/alerts/resolved`, ({ request }) => {
 		return mockAlertsList(request, playgroundState.resolvedAlerts.map(withLastComment));
+	}),
+
+	// The Insights page: the REAL analytics engine (shared with the server) over
+	// deterministic episode histories derived from the mock alerts, so every chart
+	// and metric behaves exactly like production.
+	http.get(`${API_BASE}/alerts/analytics`, ({ request }) => {
+		const url = new URL(request.url);
+		const from = url.searchParams.get('from');
+		const to = url.searchParams.get('to') ?? nowIso();
+		const timeZone = url.searchParams.get('tz') ?? undefined;
+		const tagKey = url.searchParams.get('tagKey') ?? undefined;
+		const activeAlerts = playgroundState.alerts.map(withAppliedEnrichments);
+		const resolvedAlerts = playgroundState.resolvedAlerts;
+
+		// Dashboard scoping: same filter/search semantics as the list endpoints.
+		let allowedAlertIds: Set<string> | undefined;
+		try {
+			const filters = JSON.parse(url.searchParams.get('filters') ?? '{}') as Record<string, string[]>;
+			const search = url.searchParams.get('search') ?? undefined;
+			if (Object.keys(filters).length > 0 || search?.trim()) {
+				const { items } = applyAlertListQuery([...activeAlerts, ...resolvedAlerts], playgroundState.users, {
+					filters,
+					search,
+				});
+				allowedAlertIds = new Set(items.map((a) => a.id));
+			}
+		} catch {
+			return HttpResponse.json({ success: false, error: 'Validation error' }, { status: 400 });
+		}
+
+		const episodes = [
+			...activeAlerts.flatMap((alert) => generateEpisodeRows(alert, true)),
+			...resolvedAlerts.flatMap((alert) => generateEpisodeRows(alert, false)),
+		];
+		return HttpResponse.json({
+			success: true,
+			data: computeAlertAnalytics({
+				episodes,
+				events: analyticsEvents([...activeAlerts, ...resolvedAlerts]),
+				activeAlerts,
+				resolvedAlerts,
+				from,
+				to,
+				timeZone,
+				allowedAlertIds,
+				tagKey,
+			}),
+		});
 	}),
 
 	http.patch(`${API_BASE}/alerts/:alertId/silence`, async ({ params, request }) => {
